requirements.txt
fastapi
uvicorn
pillow
python-multipart
aiofiles
python-telegram-bot==13.15
httpx
replicate
requests
numpy
insightface
onnxruntime
matplotlib